fix: отдельный компонент для шаблона
