+++ src/components/dashbord/templates/templates.tsx
@@ -1,0 +1,80 @@
+import { FC, useEffect, useRef, useState } from 'react';
+import { useDraggable } from 'react-use-draggable-scroll';
+import styles from './templates.module.scss';
+
+import { useAppDispatch, useAppSelector } from '../../../services/hooks/hooks';
+import { getTemplatesBotsAction } from '../../../services/actions/bots/templatesBots';
+import { getAccessToken } from '../../../auth/authService';
+import { getTemplatesBotsSel } from '../../../utils/selectorData';
+import Typography from '../../../ui/typography/typography';
+import Template from '../template/template';
+
+const Templates: FC = () => {
+  const { templatesBots } = useAppSelector(getTemplatesBotsSel);
+
+  const dispatch = useAppDispatch();
+
+  const token = getAccessToken();
+
+  useEffect(() => {
+    dispatch(getTemplatesBotsAction(token));
+  }, [dispatch]);
+
+  const [isExpanded, setIsExpanded] = useState(false);
+
+  const ref =
+    useRef<HTMLDivElement>() as unknown as React.MutableRefObject<HTMLUListElement>;
+  const { events } = useDraggable(ref);
+
+  return (
+    <div className={styles.container}>
+      <div className={styles.header_wrapper}>
+        <Typography tag="h2" fontFamily="secondary">
+          Шаблоны
+        </Typography>
+        <div className={styles.accordion_wrapper}>
+          <button
+            className={styles.accordion_button}
+            type="button"
+            onClick={() => {
+              setIsExpanded(!isExpanded);
+            }}
+            aria-label="xs"
+          >
+            Все шаблоны
+          </button>
+
+          <button
+            className={`${styles.accordion_ico_plus} ${
+              isExpanded ? styles.accordion_ico_minus : ''
+            }`}
+            type="button"
+            onClick={() => {
+              setIsExpanded(!isExpanded);
+            }}
+            aria-label="xs"
+          />
+        </div>
+      </div>
+      <ul
+        className={`
+          ${styles.list} ${
+            isExpanded ? styles.expanded : styles.not_expanded
+          } `}
+        {...events}
+        ref={ref}
+      >
+        {templatesBots !== null &&
+          templatesBots.map((templateBot) => (
+            <Template
+              // eslint-disable-next-line no-underscore-dangle
+              key={templateBot._id}
+              bot={templateBot}
+            />
+          ))}
+      </ul>
+    </div>
+  );
+};
+
+export default Templates;
